+++ config/routes.rb
@@ -26,4 +26,18 @@ RailsConsoleAi::Engine.routes.draw do
       end
     end
   end
+
+  resources :agents do
+    member do
+      post :approve
+    end
+    collection do
+      get :diff
+    end
+    resources :versions, only: [:index, :show], controller: 'agent_versions' do
+      member do
+        post :restore
+      end
+    end
+  end
 end